agents/agent-job-run/pyproject.toml
[project]
name = "agent-job-run"
version = "0.1.0"
description = "Demo: Async agentic workflows with Databricks Lakeflow Jobs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "databricks-sdk>=0.38.0",
    "databricks-mcp>=0.3.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]
